feat: Added the User context and basic structure for future contexts
Four main files are necessary for each context: action-types,
initial-state, reducer, and context. The action-types and
initial-state files are used to define constants and data types
for use in the reducer and context files. The reducer sets up the
initial state for the reducer and define the handlers for each of
action types. The context is used to provide access to the reducer
in the different parts of the application that needs it.